Assignment 1 | Project Layered pattern 적용
