+++ teamB/kimSungHoon/.env.sample
@@ -1,4 +1,4 @@
-DATABASE_URL="mysql://root:kambo@127.0.0.1:3306/database"
+DATABASE_URL="type://username:password@127.0.0.1:3306/database"
 
 PORT = your port number should be here !
