+++ teamB/kimSungHoon/app.js
@@ -30,33 +30,3 @@ const start = async () => {
 };
 
 start();
-
-
-// app.get('/posts-from-user/:userid', async (req,res)=>{
-
-//     const userid  = parseInt(req.params.userid);
-
-//     await Promise.all([
-//         appDataSource.query(
-//             `SELECT
-//                 posts.id AS postingId,
-//                 posts.title AS postingTitle,
-//                 posts.content AS postingContent
-//             FROM posts
-//             WHERE user_id= ${userid};
-//             `, 
-//         ),
-//         appDataSource.query(
-//             `SELECT
-//                 users.id AS userId,
-//                 users.profile_image AS userProfileImage
-//             FROM users
-//             WHERE id = ${userid};
-//             `
-//         )])
-//         .then(([postingsResult, userResult])=>{
-//             userResult[0]['postings'] = postingsResult;
-//             res.status(200).json({ "data" : userResult[0] });
-//         })
-    
-// })
