+++ teamB/kimSungHoon/controllers/postController.js
@@ -17,6 +17,18 @@ const createPost = async (req, res) => {
     }
 };
 
+const readingPost = async (req, res) => {
+    try {
+        const result = await postService.readingPost();
+        return res.status(201).json({ data: result });
+    }
+    catch(err){
+        console.log(err);
+        return res.status(err.statusCode || 500).json({message: err.message});
+    }
+};
+
 module.exports = {
-	createPost
+	createPost,
+    readingPost
 }
