+++ teamB/kimSungHoon/controllers/userController.js
@@ -17,6 +17,19 @@ const signUp = async (req, res) => {
     }
 };
 
+const userPosting = async (req, res) => {
+    try {
+        const id = req.params.id;
+        const result = await userService.userPosting(id);
+        return res.status(201).json({ data: result });
+    }
+    catch(err){
+        console.log(err);
+        return res.status(err.statusCode || 500).json({message: err.message});
+    }
+}
+
 module.exports = {
-	signUp
+	signUp,
+    userPosting
 }
