--- teamB/kimSungHoon/routes/postAllRouter.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const postAllController = require('../controllers/postAllController');
-
-const router = express.Router();
-
-router.get('/reading', postAllController.readingPost);
-
-module.exports = {
-    router 
-};
