+++ teamB/kimSungHoon/routes/postRouter.js
@@ -5,6 +5,8 @@ const router = express.Router();
 
 router.post('/register', postController.createPost);
 
+router.get('/allReading', postController.readingPost);
+
 module.exports = {
     router 
 };
