+++ teamB/kimSungHoon/routes/userRouter.js
@@ -5,6 +5,8 @@ const router = express.Router();
 
 router.post('/signup', userController.signUp);
 
+router.get('/userposting/:id', userController.userPosting);
+
 module.exports = {
     router 
 };
