--- teamB/kimSungHoon/services/postAllService.js
@@ -1,11 +1,0 @@
-const postAllDao = require('../models/postAllDao');
-
-const readingPost = async () => {
-    const result = await postAllDao.readingPost();
-    return result;
- };
-
-
-module.exports = {
-    readingPost
-}
